LayoutTests/http/tests/navigation/target-blank-opener-post.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<p>Test that rel="noreferrer" is supported by form elements.</p>
<form id="testForm" action="resources/target-blank-opener-post-window.py" target="_blank" rel="noreferrer" method="post">
<input type="submit" value="Test"></input>
<form>
<script>
if (window.testRunner)
    testRunner.waitUntilDone();

document.forms[0].submit();

</script>
</body>
</html>
