LayoutTests/http/tests/navigation/target-blank-opener-post-expected.txt

CONSOLE MESSAGE: Referrer: PASS
CONSOLE MESSAGE: window.opener: PASS
Test that rel="noreferrer" is supported by form elements.
